requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-appconf==1.0.5
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
pilkit==2.0
Pillow==9.1.0
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
zipp==3.8.0
